waseda-grade-api/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
python-multipart
selenium
webdriver-manager
mysql-connector-python
